% Generated by roxygen2 (4.0.1): do not edit by hand
\name{traitNames}
\alias{assemblage.phylogenies}
\alias{envNames}
\alias{sites}
\alias{sites<-}
\alias{species}
\alias{species<-}
\alias{traitNames}
\title{Describing and manipulating comparativbe.comm objects}
\usage{
traitNames(object)

envNames(object)

species(x)

species(x) <- value

sites(x)

sites(x) <- value

assemblage.phylogenies(data)
}
\arguments{
\item{object}{A \code{\link{comparative.comm}} object}
}
\value{
Names of the traits or environmental variables
}
\description{
Describing and manipulating comparativbe.comm objects
}

