% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cc.traitgram}
\alias{cc.traitgram}
\alias{princomp.one}
\title{Traitgram for comparative community object}
\usage{
cc.traitgram(object, trait, moreArgs = NULL, ...)

\method{princomp}{one}(x, ...)
}
\arguments{
\item{object}{A \code{\link{comparative.comm}} object.}

\item{trait}{Which trait to plot.  If \code{\link{missing}}, use
the first trait.  If a positive \code{\link{numeric}} vector of
\code{\link{length}} one, use the \code{as.integer(trait)}th
trait.  If a \code{\link{numeric}} vector, use it instead of the
trait data in \code{object}.  If a \code{\link{character}} vector
of \code{\link{length}} one, use the trait with that name.  If a
\code{\link{function}} pass the trait data frame through that
function and use the result (\code{\link{princomp.one}} is a
wrapper).  If an \code{\link{expression}}, evaluate that
expression in the environment of the trait data and use the
result.  If a \code{\link{character}} vector, then convert to an
expression and evaluate in the environment of the trait data and
use the result.}

\item{moreArgs}{List of more arguments to pass on to \code{trait}
(if its a \code{\link{function}}).}

\item{...}{Additional arguments to be passed on to
\code{\link{traitgram}}.}

\item{x}{A matrix-like object}

\item{...}{Arguments to pass on to \code{\link{princomp}}}
}
\value{
See \code{\link{traitgram}}

The first axis of a PCA
}
\description{
A wrapper for the \code{\link{traitgram}} function in the
\code{picante} package.

A very soft wrapper for \code{\link{princomp}}
}

