% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coef.phy.structure}
\alias{coef.phy.structure}
\title{Get coefs of a phylogenetic structure object}
\usage{
\method{coef}{phy.structure}(object, ...)
}
\arguments{
\item{object}{\code{phy.structure} object with coefficients to be
returned}

\item{...}{(ignored)}
}
\description{
Get coefs of a phylogenetic structure object
}
\note{
Calling \code{object$coefs} would give you this too!
}

