% Generated by roxygen2 (4.0.1): do not edit by hand
\name{comm.dist}
\alias{comm.dist}
\alias{comm.dist.comparative.comm}
\alias{comm.dist.matrix}
\title{Make ecological co-existence matrices}
\usage{
comm.dist(x)

\method{comm.dist}{matrix}(x)

\method{comm.dist}{comparative.comm}(x)
}
\arguments{
\item{x}{an object}
}
\description{
Make ecological co-existence matrices
}
\details{
Returns the 1 - co-existence of species. Look at how this
is calcualted; it incorporates abundances, and if you don't want it
to do so simply call it on a presence/absensence (1/0) matrix.
}
\seealso{
Other distances: \code{\link{dist.func.default}},
  \code{\link{traits.dist}},
  \code{\link{traits.dist.comparative.comm}},
  \code{\link{traits.dist.data.frame}},
  \code{\link{traits.dist.default}};
  \code{\link{funct.phylo.dist}}; \code{\link{phylo.dist}},
  \code{\link{phylo.dist.comparative.comm}},
  \code{\link{phylo.dist.phylo}};
  \code{\link{pianka.dist}},
  \code{\link{pianka.dist.comparative.comm}},
  \code{\link{pianka.dist.matrix}}
}

