% Generated by roxygen2 (4.0.1): do not edit by hand
\name{comparative.comm}
\alias{comparative.comm}
\alias{print.comparative.comm}
\title{\code{comparative.comm} creates a community comparative ecology object}
\usage{
comparative.comm(phy, comm, traits = NULL, env = NULL, warn = TRUE,
  vcv = FALSE, force.root = -1)

\method{print}{comparative.comm}(x, ...)
}
\arguments{
\item{phy}{phylogeny (in ape::phylo format) of species}

\item{comm}{a community matrix (as used in vegan) with species as columns and rows as community samples}

\item{traits}{a data.frame of species traits (with row names matching those of 'comm')}

\item{env}{a data.frame of environmental data (all continuous
variables) with row names matching those of 'comm'}

\item{warn}{whether to warn if species/sites are dropped when
creating object (default: TRUE)}

\item{vcv}{whether to calculate variance-covariance matrix and
store in object to save time (but perhaps not memory) for later
calculations (default: FALSE). Storing a VCV now may cause problems
if you intend to use this object with the package \code{caper}.}

\item{force.root}{if phylogeny is unrooted, a root.edge of value
force.root will be added (default: -1, which means this will never
happen). Rarely needed, even more rarely advisable.}

\item{x}{\code{comparative.comm} object to be printed}

\item{...}{not currently used}
}
\value{
comparative.comm object
}
\description{
\code{comparative.comm} creates a community comparative ecology object
}
\details{
Basic checking of whether the input data match up is
performed; you need only supply 'comm' and 'phy', nothing else is
mandatory.  Subsetting according to communities (rows) and species
(columns) is possible - see 'examples' for details.
}
\examples{
\dontrun{
data(laja)
data <- comparative.comm(invert.tree, river.sites)
data <- comparative.comm(invert.tree, river.sites, invert.traits)
data <- comparative.comm(invert.tree, river.sites, invert.traits, river.env)
data[1:3,]
data[,1:3]
}
}

