% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dissimilarity}
\alias{dissimilarity}
\title{Calculate dissimilarity phylogenetic biodiversity metrics across communities}
\usage{
dissimilarity(data, metric = c("all", "unifrac", "pcd", "phylosor",
  "comdist"), abundance = TRUE, permute = 100, ...)
}
\arguments{
\item{data}{a \code{comparative.comm} object}

\item{metric}{specify particular metrics to calculate, default is \code{all}}

\item{abundance}{If TRUE (default) metrics are calculated incorporating species abundances (currently only comdist)}

\item{permute}{Number of permutations for metric (currently only for PCD)}

\item{...}{additional parameters to be passed to `metric function(s) you are calling}
}
\value{
a \code{phy.structure} list object of metric values
}
\description{
\code{dissimilarity} calculates phylogenetic biodiversity metrics
}
\details{
Calculates various metrics of phylogenetic biodiversity that are categorized as \emph{dissimilarity} metrics by Pearse \emph{et al.} (2014). WARNING: Phylosor is presented as a distance matrix here, i.e. it is *not* the fraction of shared branch length among communities, but rather '1 - shared branch length'. This means \code{dissimilarity} returns a *distance* object, not a similarity object. This is different from the output of other R packages, but more convenient for us!
}
\note{
This function uses a version of the PCD function, that is not included in \code{picante} and can be slow if \code{metric}="all"
}
\examples{
\dontrun{
data(laja)
data <- comparative.comm(invert.tree, river.sites, invert.traits)
dissimilarity(data)
dissimilarity(data, "pcd")
}
}
\author{
M.R. Helmus, Will Pearse
}
\references{
Pearse W.D., Purvis A., Cavender-Bares J. & Helmus
M.R. (2014). Metrics and Models of Community Phylogenetics. In:
Modern Phylogenetic Comparative Methods and Their Application in
Evolutionary Biology. Springer Berlin Heidelberg, pp. 451-464.

\code{unifrac} Lozupone C.A. & Knight
R. (2005). UniFrac: a new phylogenetic method for comparing
microbial communities. Applied and Environmental Microbiology, 71,
8228-8235.

\code{pcd} Ives A.R. & Helmus M.R. (2010). Phylogenetic
metrics of community similarity. The American Naturalist, 176,
E128-E142.

\code{phylosor} Bryant J.A., Lamanna C., Morlon H.,
Kerkhoff A.J., Enquist B.J. & Green J.L. (2008). Microbes on
mountainsides: Contrasting elevational patterns of bacterial and
plant diversity. Proceedings of the National Academy of Sciences of
the United States of America, 105, 11505-11511.

\code{comdist} C.O. Webb, D.D. Ackerly, and
S.W. Kembel. 2008. Phylocom: software for the analysis of
phylogenetic community structure and trait
evolution. Bioinformatics 18:2098-2100.
}

