% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drop_tip}
\alias{drop_tip}
\title{Trim a phylogeny}
\usage{
drop_tip(tree, spp)
}
\arguments{
\item{tree}{A \code{phylo} object}

\item{spp}{A tip to drop}
}
\value{
a no-nonsense phylogeny
}
\description{
This is an \code{ape} work-around, intended to be the same as
\code{\link{drop.tip}} but not returning a nonsense phylogeny).
}

