% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eco.env.regression}
\alias{eco.env.regression}
\title{Run an eco.env.regression regression (Cavender-Bares et al. 2004)}
\usage{
eco.env.regression(data, randomisation = c("taxa.labels", "richness",
  "frequency", "sample.pool", "phylogeny.pool", "independentswap", "trialswap"),
  permute = 0, n.clades = 1, method = c("quantile", "lm", "mantel"),
  swap.iter = 1000, altogether = TRUE, ...)
}
\arguments{
\item{data}{a comparative community ecology object on which to run the regression}

\item{randomisation}{what kind of null distributions to compare your data with - one of:
taxa.labels, richness, frequency, sample.pool, phylogeny.pool, independentswap, trialswap
(as implemented in 'picante')}

\item{permute}{the number of null permutations to perform}

\item{altogether}{use distance matrix based on all traits (default
TRUE), or perform separate regressions for each trait}

\item{...}{additional arguments to pass on to model fitting functions}
}
\description{
\code{eco.env.regression} runs a regression of ecological community
coexistence versus species' trait-distances' overlap based on
environmental tolerances
}
\details{
This is extremely unchcked, so beware!
}
\note{
Like the eco.trait and eco.env methods, this is a data-hungry
method. Warnings will be generated if any of the methods cannot be
fitted properly (the examples below give toy examples of this). In
such cases the summary and plot methods of these functions may
generate errors; use 'traceback()' to examine where these are
coming from, and consider whether you want to be working with the
data generating these errors. I am loathe to hide these errors or
gloss over them, because they represent the reality of your data!
}
\examples{
\dontrun{
data(phylocom)
dummy.env <- data.frame(letters=letters[1:6], repeated=rep("A", 6), row.names=rownames(phylocom$sample))
data <- comparative.comm(phylocom$phy, phylocom$sample, env=dummy.env)
eco.env.regression(data, permute=10)
plot(eco.env.regression(data, permute=10, method="lm))
plot(eco.env.regression(data, permute=10, method="lm", altogether=FALSE))
}
}
\author{
Will Pearse, Jeannine Cavender-Bares
}

