% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{eco.xxx.regression.list}
\alias{eco.xxx.regression.list}
\alias{plot.eco.xxx.regression.list}
\alias{print.eco.xxx.regression.list}
\alias{summary.eco.xxx.regression.list}
\title{List of eco.xxx.regressions}
\usage{
\method{summary}{eco.xxx.regression.list}(object, ...)

\method{print}{eco.xxx.regression.list}(x, ...)

\method{plot}{eco.xxx.regression.list}(x, ...)
}
\arguments{
\item{object}{\code{eco.xxx.regression.list} object}

\item{...}{additional arguments to plotting functions}

\item{x}{\code{eco.xxx.regression.list} object}
}
\description{
List of eco.xxx.regressions
}

