\docType{package}
\name{EcoPhyl-package}
\alias{EcoPhyl}
\alias{EcoPhyl-package}
\title{Eco-Phylognetic Analysis (in R)}
\description{
  Eco-Phylognetic Analysis (in R)
}
\details{
  \tabular{ll}{ Package: \tab EcoPhyl\cr Type: \tab
  Package\cr Version: \tab 0.2\cr Date: \tab 2013-10-15\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }

  Eco-phylogenetic analysis; regressions of community
  co-existence against phylogenetic dissimilarity and niche
  overlap
}
\author{
  William D. Pearse \email{wdpearse@umn.edu}

  Matthew R. Helmus
}

