% Generated by roxygen2 (4.0.1): do not edit by hand
\name{edge2phylo}
\alias{edge2phylo}
\title{\code{edge2phylo} convert a simulated edge matrix to an ape:phylo aobject}
\usage{
edge2phylo(edge, species, extinct = numeric(0), edge.length = NA,
  traits = NULL)
}
\arguments{
\item{edge}{a two-column matrix where the first column is the start
node, the second the destination}

\item{species}{which of the rows in the edge matrix are extant
species}

\item{extinct}{which of the tips in the edge matrix are extinct}

\item{edge.length}{if given (default NA), a vector to be used to
give edge.length to the phylogeny}

\item{traits}{if given (default NA), a vector to be used for traits
($traits) in the phylogeny}
}
\value{
ape::phylo object with random tip.labels
}
\description{
\code{edge2phylo} convert a simulated edge matrix to an ape:phylo aobject
}
\details{
This is a bit OTT because you can figure out what's
extinct and extant from the edge matrix itself; I'm just lazy and
this matches onto exactly what my code already does. See
sim.bd.tree for use; it's really an internal function.
}
\author{
Will Pearse
}

