% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fingerprint.regression}
\alias{fingerprint.regression}
\title{Plot a fingerprint regression (sensu Cavender-Bares et al. 2004 Figure 6)}
\usage{
fingerprint.regression(data, eco.rnd = c("taxa.labels", "richness",
  "frequency", "sample.pool", "phylogeny.pool", "independentswap", "trialswap"),
  eco.method = c("quantile", "lm", "mantel"), eco.permute = 1000,
  eco.swap = 1000, evo.method = c("lambda", "delta", "kappa"),
  evo.permute = 1000, ...)
}
\arguments{
\item{data}{a comparative community ecology object on which to run the analyses}

\item{eco.rnd}{null distribution with which to compare your community data - one of:
taxa.labels, richness, frequency, sample.pool, phylogeny.pool, independentswap, trialswap
(as implemented in 'picante')}

\item{eco.method}{regression to perform - one of: lm, quantile, mantel}

\item{eco.permute}{the number of null permutations to perform}

\item{eco.swap}{the number of independent swaps to perform (if doing so)}

\item{evo.method}{how to measure phylogenetic inertia - one of: lambda, delta, kappa, D.c}

\item{evo.permute}{number of permutations of phylogenetic inertia (D.c only)}

\item{...}{additional parameters to pass on to model fitting functions}
}
\description{
\code{fingerprint.regression} calculates traits' phylogenetic
inertia, and then plots this as a function of trait similarity
among communities
}
\details{
This is extremely unchcked, so beware!
}
\note{
Like the eco.trait and eco.env methods, this is a data-hungry
method. Warnings will be generated if any of the methods cannot be
fitted properly (the examples below give toy examples of this). In
such cases the summary and plot methods of these functions may
generate errors; use 'traceback()' to examine where these are
coming from, and consider whether you want to be working with the
data generating these errors. I am loathe to hide these errors or
gloss over them, because they represent the reality of your data!
}
\examples{
\dontrun{
data(phylocom)
data <- comparative.comm(phylocom$phy, phylocom$sample, traits=phylocom$traits)
fingerprint.regression(data, eco.permute=10)
plot(fingerprint.regression(data, permute=10, method="lm))
plot(fingerprint.regression(data, permute=10, method="lm", altogether=FALSE))
}
}
\author{
Will Pearse
}

