% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FPDist}
\alias{FPDist}
\title{Functional phylogenetic distance}
\usage{
FPDist(data, a = 0.5, p = 1)
}
\arguments{
\item{data}{comparative.comm object with traits and a phylogeny}

\item{a}{A number between 0 and 1 giving the amount of weight.
to put on \code{PDist} relative to \code{FDist}.}

\item{p}{A number giving the \code{p}-norm.}
}
\value{
A distance matrix.
}
\description{
Computates a functional phylogenetic distance matrix from
phylogenetic and functional distance matrices and two
weighting parameters.
}
\note{
This function is not very user friendly yet.  There are
	no doubt many use cases that I've ignored.
}

