% Generated by roxygen2 (4.0.1): do not edit by hand
\name{funct.phylo.dist}
\alias{funct.phylo.dist}
\title{Make functional phylogenetic distance matrix}
\usage{
funct.phylo.dist(x, phyloWeight, p, ...)
}
\arguments{
\item{x}{a \code{\link{comparative.comm}} object}

\item{phyloWeight}{phylogenetic weighting parameter (referred to as
\code{a} in Cadotte et al. (2013)}

\item{p}{exponent giving the exponent to use for combining
functional and phylogenetic distances (\code{p = 2} gives a
Euclidean combination).}

\item{...}{not currently used}
}
\description{
Make functional phylogenetic distance matrix
}
\seealso{
Other distances: \code{\link{comm.dist}},
  \code{\link{comm.dist.comparative.comm}},
  \code{\link{comm.dist.matrix}};
  \code{\link{dist.func.default}},
  \code{\link{traits.dist}},
  \code{\link{traits.dist.comparative.comm}},
  \code{\link{traits.dist.data.frame}},
  \code{\link{traits.dist.default}};
  \code{\link{phylo.dist}},
  \code{\link{phylo.dist.comparative.comm}},
  \code{\link{phylo.dist.phylo}};
  \code{\link{pianka.dist}},
  \code{\link{pianka.dist.comparative.comm}},
  \code{\link{pianka.dist.matrix}}
}

