% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metric.null.R
\name{generic.metric}
\alias{.metric.null}
\alias{.ses}
\alias{generic.metric}
\alias{generic.null}
\title{Compare any metric(s) with null distributions}
\usage{
generic.null(data, metrics, null.model = c("taxa.labels", "richness",
  "frequency", "sample.pool", "phylogeny.pool", "independentswap", "trialswap"),
  permute = 1000, comp.fun = .ses, ...)

.ses(observed, null)

.metric.null(data, metrics, null.model = c("taxa.labels", "richness",
  "frequency", "sample.pool", "phylogeny.pool", "independentswap", "trialswap"),
  permute = 1000, ...)

generic.metric(data, metrics, ...)
}
\arguments{
\item{data}{data \code{\link{comparative.comm}} object}

\item{metrics}{vector of functions to be calculated on \code{data};
see \link{pez.metrics} for a list of them.}

\item{null.model}{one of "taxa.labels", "richness", "frequency",
"sample.pool", "phylogeny.pool", "independentswap", or
"independentswap". These correspond to the null models available in
\code{\link{picante}}}

\item{permute}{number of null permutations to perform (default
1000)}

\item{comp.fun}{comparison function to compare observed values with
null values. Default is \code{\link{.ses}}; this is a Standard
Effect Size (obs - mean)/SEmean. You may supply your own function;
it should take the observed site-metric matrix as its first
argument, and a site-metric-permutation array as its second. See
the internals of \code{\link{generic.null}} for an example of its
use.}

\item{...}{additional arguments (e.g, \code{dist},
\code{abundance.weighted}) to be passed to any metric functions
(see \code{\link{generic.metrics}} for possible arguments)}

\item{observed}{observed metric values in site-metric matrix
(\emph{e.g.}, from \link{\code{generic.metrics}})}

\item{null}{null distributions (\emph{e.g.}, from
\link{\code{.metric.null}}) in a site-metric-permutation array}

\item{permute}{number of null permutations}
}
\value{
Output from \code{comp.fun}, by default a vector of
standard effect sizes

Vector of standard effect sizes

site-metric-permutation array

site-metric matrix
}
\description{
Compare any calculated metric within \code{pez} to values expected
under some null distribution. This is a very light wrapper around
the utility functions \code{\link{.generic.null}} and
\code{\link{generic.metrics}} (which is, itself, a very simple
function!).

Calculate Standard Effect Sizes of metrics

Produce null randomisations and compute metrics across them

Calculate arbitrary metrics across data
}
\note{
\code{comp.fun} can be \emph{anything}; much ink has been
written about the use of standard effect sizes in eco-phylogenetic
analyses (\emph{e.g.}, Kembel 2009). That this function makes it
easy for you to compute Standard Effect Sizes does not necessarily
mean that you should (see Pearse et al. 2013).

Calculating null permutations on a dispersion metric makes little
sense, since (by definition; see Pearse et al. 2014) a dispersion
metric \emph{require} the use of a null distribution to be
calculated. There is nothing to stop you doing so, however! The
code makes no attempt to stop you calculating null dissimilarity
metrics, but I am not certain that doing so is a good idea using
this code as I don't know what to do with such null models!
}
\author{
Will Pearse
}
\references{
Kembel S.W. (2009) Disentangling niche and neutral
influences on community assembly: assessing the performance of
community phylogenetic structure tests. Ecology letters, 12(9),
949-960.

Pearse W.D., Jones F.A. & Purvis A. (2013) Barro
Colorado Island's phylogenetic assemblage structure across fine
spatial scales and among clades of different ages. Ecology, 94(12),
2861-2872.
}

