\docType{data}
\name{laja}
\alias{laja}
\title{Macroinvertebrate samples from the Rio Laja of Mexico}
\format{\code{laja} contains a \code{\link{phylo}} object, a dataframe of sites-by-taxa, and a dataframe of traits}
\usage{
laja
}
\description{
This data set includes macroinvertebrate samples from the
Rio Laja, a phylogenetic tree of the taxa and traits that
include mean body length and fish feeding preference as in
Helmus \emph{et al.} 2013.
}
\author{
M.R. Helmus
}
\references{
Helmus M., Mercado-Silva N. & Vander Zanden M.J. (2013).
Subsidies to predators, apparent competition and the
phylogenetic structure of prey communities. Oecologia, 173,
997-1007.
}
\keyword{datasets}

