\name{numnames}
\alias{numnames}
\title{Generates numbered names}
\usage{
numnames(n, prefix = "name")
}
\arguments{
  \item{n}{Number of names.}

  \item{prefix}{Character prefix to go in front of the
  numbers.}
}
\value{
A character vector of numbered names.
}
\description{
Generates numbered names based on a character prefix that
will be alphanumerically sorted as expected.
}

