% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pblm.traits}
\alias{pblm.traits}
\title{Calculate dispersion metrics across communities}
\usage{
pblm.traits(assocs, tree1 = NULL, tree2 = NULL, covars1 = NULL,
  covars2 = NULL, bootstrap = FALSE, nreps = 10, maxit = 10000,
  pstart = c(0.5, 0.5))
}
\arguments{
\item{data}{a comparative community ecology object}

\item{permute}{the number of null permutations to perform}

\item{metric}{specify (a) particular metric(s) to calculate (sesmpd, sesmntd, sespd, innd, d), or the default 'all'}
}
\value{
cc.dispersion object (a named list with the output from each metric)
}
\description{
\code{dispersion} calculates dispersion metrics in comparative.comm communities
}
\details{
This calculates $SES_{MPD}$, $SES_{MNTD}$, $SES_{PD}$, INND, and D. All these are defined as dispersion metrics in Pearse et al.,
Dc=0 is the Brownian expectation, Dc=1 is the random expectation.
}
\examples{
\dontrun{
data(phylocom)
data <- comparative.comm(phylocom$phy, phylocom$sample)
dispersion(data)
dispersion(data, 100, "sesmpd")
}
}
\author{
Matt Helmus, Will Pearse
}

