\name{pez-internal}
\alias{.Random.seed}
\alias{.eco.evo.clade.regression}
\alias{.eco.evo.regression}
\alias{.eco.null}
\alias{.eco.phy.regression}
\alias{.plot.regression}
\alias{.prepare.regression.output}
\alias{.summary.regression}
\title{Internal pez Functions}
\description{
 Internal Grid functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :p).
}
\keyword{internal}