% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{pez.metrics}
\alias{.aed}
\alias{.colless}
\alias{.comdist}
\alias{.d}
\alias{.delta}
\alias{.dist.fd}
\alias{.eaed}
\alias{.eed}
\alias{.eigen.sum}
\alias{.gamma}
\alias{.haed}
\alias{.hed}
\alias{.iac}
\alias{.inmd}
\alias{.innd}
\alias{.kappa}
\alias{.lambda}
\alias{.mntd}
\alias{.mpd}
\alias{.pae}
\alias{.pcd}
\alias{.pd}
\alias{.phylo.entropy}
\alias{.phylosor}
\alias{.pse}
\alias{.psr}
\alias{.psv}
\alias{.rao}
\alias{.scheiner}
\alias{.ses.mntd}
\alias{.ses.mpd}
\alias{.simpson.phylogenetic}
\alias{.sqrt.phy}
\alias{.taxon}
\alias{.unifrac}
\alias{pez.metrics}
\title{Colless' index (Colless, 1982)}
\usage{
.colless(data, ...)

.hed(data, ...)

.eed(data, na.rm = TRUE, ...)

.psv(x, ...)

.psr(x, ...)

.mpd(x, dist = NULL, abundance.weighted = FALSE, ...)

.pd(x, include.root = TRUE, ...)

.mntd(x, dist = NULL, abundance.weighted = FALSE, ...)

.gamma(x, ...)

.taxon(x, dist = NULL, ...)

.eigen.sum(x, dist = NULL, which.eigen = 1, ...)

.dist.fd(x, method = "phy", abundance.weighted = FALSE, ...)

.sqrt.phy(x)

.phylo.entropy(data, ...)

.aed(data, ...)

.haed(data, ...)

.simpson.phylogenetic(data)

.iac(data, na.rm = TRUE, ...)

.pae(data, na.rm = TRUE, ...)

.scheiner(data, q = 0, abundance.weighted = TRUE, ...)

.pse(x, ...)

.rao(x, ...)

.lambda(x, ...)

.delta(x, ...)

.kappa(x, ...)

.eaed(x, ...)

.unifrac(x, ...)

.pcd(x, permute = 1000, ...)

.comdist(x, dist = NULL, abundance.weighted = FALSE, ...)

.phylosor(x, dist = NULL, abundance.weighted = FALSE, ...)

.d(data, permute = 1000, ...)

.ses.mpd(x, dist = NULL, null.model = "taxa.labels",
  abundance.weighted = FALSE, permute = 1000, ...)

.ses.mntd(x, dist = NULL, null.model = "taxa.labels",
  abundance.weighted = FALSE, permute = 1000, ...)

.inmd(x, dist = NULL, null.model = "taxa.labels",
  abundance.weighted = FALSE, permute = 1000, ...)

.innd(x, dist = NULL, null.model = "taxa.labels",
  abundance.weighted = FALSE, permute = 1000, ...)
}
\description{
Colless' index (Colless, 1982)

Hed (Cadotte et al., 2010)

Eed (Cadotte et al., 2010)

PSV (Helmus et al., 2007)

PSR (Helmus et al., 2007)

MPD (Mean Phylogenetic Distance)

PD (Faith, 1992)

MNTD (Mean Nearest Taxonomic Distance)

Gamma (Pybus & Harvey, 2000)

Taxonomic distinctiveness (Clarke & Warwick, 1998)

Sum of Phylo-Eigvenectors (Diniz-Filho et al., 2011)

Functional Diversity

Phylogenetic entropy (Allen et al., 2009)

XXX

Haed (Cadotte et al., 2010)

Simpson's phylogenetic diversity

IAC (Cadotte et al., 2010)

PAE (Cadotte et al., 2010)

PSE (Helmus et al., 2007)

Rao's Quadratic Entropy (XXX)

Pagel's Lambda (Pagel, 1999)

Pagel's Delta (Pagel, 1999)

Pagel's Kappa (Pagel, 1999)

Eaed (Cadotte et al., 2010)

UniFrac (Lozupone et al., 2005)

PCD (Ives & Helmus, 2010)

ComDist (Webb et al., 2008)

PhyloSor (Bryant et al., 2008)

D (Fritz & Purvis, 2010)

SESmpd (Webb, 2000)

SESmntd (Webb, 2000)

INMD (XXX)

INND (XXX)
}
\references{
\code{colless} Colless D.H. (1982). Review of
phylogenetics: the theory and practice of phylogenetic
systematics. Systematic Zoology, 31, 100-104.

\code{eed,hed} (i.e., \emph{Eed, Hed}) Cadotte M.W.,
Davies T.J., Regetz J., Kembel S.W., Cleland E. & Oakley
T.H. (2010). Phylogenetic diversity metrics for ecological
communities: integrating species richness, abundance and
evolutionary history. Ecology Letters, 13, 96-105.

\code{PSV,PSR,PSE} Helmus M.R., Bland T.J., Williams
C.K. & Ives A.R. (2007). Phylogenetic measures of
biodiversity. American Naturalist, 169, E68-E83.

\code{PD} Faith D.P. (1992). Conservation evaluation
and phylogenetic diversity. Biological Conservation, 61, 1-10.

\code{gamma} Pybus O.G. & Harvey P.H. (2000) Testing
macro-evolutionary models using incomplete molecular
phylogenies. _Proceedings of the Royal Society of London. Series
B. Biological Sciences 267: 2267--2272.

\code{taxon} Clarke K.R. & Warwick R.M. (1998). A
taxonomic distinctness index and its statistical
properties. J. Appl. Ecol., 35, 523-531.

\code{eigen.sum} Diniz-Filho J.A.F., Cianciaruso M.V.,
Rangel T.F. & Bini L.M. (2011). Eigenvector estimation of
phylogenetic and functional diversity. Functional Ecology, 25,
735-744.

\code{entropy} Allen B., Kon M. & Bar-Yam Y. (2009). A
New Phylogenetic Diversity Measure Generalizing the Shannon Index
and Its Application to Phyllostomid Bats. The American Naturalist,
174, 236-243.

\code{pae,iac,haed,eaed} Cadotte M.W., Davies T.J.,
Regetz J., Kembel S.W., Cleland E. & Oakley
T.H. (2010). Phylogenetic diversity metrics for ecological
communities: integrating species richness, abundance and
evolutionary history. Ecology Letters, 13, 96-105.

\code{rao} Webb C.O. (2000). Exploring the phylogenetic
structure of ecological communities: An example for rain forest
trees. American Naturalist, 156, 145-155.

\code{lambda,delta,kappa} Mark Pagel (1999) Inferring
the historical patterns of biological evolution. Nature 6756(401):
877--884.

\code{unifrac} Lozupone C.A. & Knight
R. (2005). UniFrac: a new phylogenetic method for comparing
microbial communities. Applied and Environmental Microbiology, 71,
8228-8235.

\code{pcd} Ives A.R. & Helmus M.R. (2010). Phylogenetic
metrics of community similarity. The American Naturalist, 176,
E128-E142.

\code{comdist} C.O. Webb, D.D. Ackerly, and
S.W. Kembel. 2008. Phylocom: software for the analysis of
phylogenetic community structure and trait
evolution. Bioinformatics 18:2098-2100.

\code{phylosor} Bryant J.A., Lamanna C., Morlon H.,
Kerkhoff A.J., Enquist B.J. & Green J.L. (2008). Microbes on
mountainsides: Contrasting elevational patterns of bacterial and
plant diversity. Proceedings of the National Academy of Sciences of
the United States of America, 105, 11505-11511.

\code{d} Fritz S.A. & Purvis A. (2010). Selectivity in
Mammalian Extinction Risk and Threat Types: a New Measure of
Phylogenetic Signal Strength in Binary Traits. Conservation
Biology, 24, 1042-1051.

\code{sesmpd,sesmntd} Webb C.O. (2000). Exploring the
phylogenetic structure of ecological communities: An example for
rain forest trees. American Naturalist, 156, 145-155.

\code{innd} Ness J.H., Rollinson E.J. & Whitney
K.D. (2011). Phylogenetic distance can predict susceptibility to
attack by natural enemies. Oikos, 120, 1327-1334.
}

