% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{pez}
\alias{package-pez}
\alias{pez}
\alias{pez-package}
\title{Phylogenetics for the Environmental Sciences}
\description{
Eco-phylogenetic and community phylogenetic analyses.  Keeps
community ecological and phylogenetic data matched up and
comparable using 'comparative.comm' objects. Wrappers for common
community phylogenetic indices ('shape', 'evenness', 'dispersion',
and 'dissimilarity' metrics). Implementation of Cavender-Bares
(2004) correlation of phylogenetic and ecological matrices
('fingerprint.regression'). Simulation of null assemblages, traits,
and phylogenies ('scape', 'sim.meta.comm').
}

