% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phy.build}
\alias{bind.replace}
\alias{merge.phylo}
\alias{phy.build}
\title{Bind missing species into a phylogeny based on taxonomy}
\usage{
merge.phylo(tree, species, split = "_")

bind.replace(backbone, donor, replacing.tip.label, donor.length = NA)
}
\arguments{
\item{tree}{\code{\link{ape::phylo}} phylogeny to have those
species inserted into it}

\item{lookup}{vector of species names to be bound into the tree if
missing from it}

\item{split}{the character that splits genus and species names in
your phylogeny. Default is \code{_}, i.e. Quercus_robur.}

\item{backbone}{backbone phylogeny (\code{\link{ape::phylo}}) into
which the donor is to be bound}

\item{donor}{phylogeny (\code{\link{ape::phylo}}) to bound into the
backbone phylogeny}

\item{replacing.tip.label}{the species in the donor phylogeny
that's being replaced by the donor phylogeny}

\item{donor.length}{how deep the donor phylogeny should be cut into
the backbone phylogeny. If NA (default), then the bladj algorithm
is followed (or, in plain English, it's put half-way along the
branch)}
}
\value{
\code{\link{ape::phylo}} phylogeny

phylogeny (\code{\link{ape::phylo}})
}
\description{
Bind missing species into a phylogeny based on taxonomy

Bind one phylogeny into another, giving an ultrametric phylogeny
}
\details{
Binds missing species into a phylogeny by replacing all
members of the clade it belongs to with a polytomy. Assumes the
\code{tip.labels} represent Latin binomials, split by the
\code{split} argument. This code was originally shipped with
phyloGenerator - this is the \code{merge} method in that program.

Binds a phylogeny (donor) into a bigger phylogeny
('backbone'); useful if you're building a phylogeny a la
Phylomatic. This code was originally shipped with phyloGenerator -
this is the \code{merge} method in that program.
}
\author{
Will Pearse

Will Pearse
}
\references{
Pearse W.D. & Purvis A. phyloGenerator: an automated phylogeny generation tool for ecologists. Methods in Ecology and Evolution 4(7): 692--698.

Pearse W.D. & Purvis A. phyloGenerator: an automated phylogeny generation tool for ecologists. Methods in Ecology and Evolution 4(7): 692--698.
}

