% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phylo.dist}
\alias{phylo.dist}
\title{Make phylogenetic distance matrices
TODO: not sure if this is necessary/useful yet}
\usage{
phylo.dist(x, ...)
}
\description{
Make phylogenetic distance matrices
TODO: not sure if this is necessary/useful yet
}

