\name{pianka.dist}
\alias{pianka.dist}
\alias{pianka.dist.comparative.comm}
\title{Make environmental tolerance distance matrices}
\usage{
pianka.dist(x, ...)

\method{pianka.dist}{comparative.comm}(x, alltogether = TRUE)
}
\arguments{
  \item{x}{an object}
}
\description{
Make environmental tolerance distance matrices
}

