% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.comparative.comm}
\alias{plot.comparative.comm}
\title{Plots dot-plots of community presence/absence or abundance}
\usage{
\method{plot}{comparative.comm}(x, sites = NULL, abundance = FALSE,
  dot.cex = NULL, site.col = "black", fraction = 3, x.increment = NULL,
  ...)
}
\arguments{
\item{x}{a comparative community ecology object (\code{comparative.comm})}

\item{sites}{names of sites to plot (default: all); see examples}

\item{abundance}{make size proportional to species abundance (default: FALSE)}

\item{dot.cex}{function to determine point size; see examples, this isn't as terrible-sounding as it seems.}

\item{site.col}{colours to use when plotting sites; if not same length as number of sites, only the first element is used (no recycling)}

\item{fraction}{fraction of plot window to be taken up with phylogeny; e.g., 3 (default) means phylogeny is 1/3 of plot}

\item{x.increment}{specify exact spacing of points along plot; see examples}

\item{...}{additional arguments passed to plotting functions}
}
\value{
List containing plot.phylo information, as well as the used x.adj values
}
\description{
Plots dot-plots of community presence/absence or abundance
}
\details{
Getting the right spacing of dots on the phylogeny may take some playing around with the fraction and x.increment arguments, see below. It may seem a little strange to set point size using a function; however, this gives you much more flexibility when dealing with abundance data, and allows you to transform the data - see the examples
}
\author{
Will Pearse, Matt Helmus
}

