\name{plot.eco.trait.regression}
\alias{plot.eco.trait.regression}
\title{\code{plot.eco.trait.regression} plots an eco.trait.regression regression}
\usage{
  \method{plot}{eco.trait.regression} (x, ...)
}
\description{
  \code{plot.eco.trait.regression} plots an
  eco.trait.regression regression
}

