% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.fingerprint.regression}
\alias{plot.fingerprint.regression}
\title{Plot a fingerprint.regression}
\usage{
\method{plot}{fingerprint.regression}(x, eco = c("slope", "corrected"),
  xlab = "Ecological Trait Coexistence", ylab = "Phylogenetic inertia", ...)
}
\description{
Plot a fingerprint.regression
}

