% Generated by roxygen2 (4.0.1): do not edit by hand
\name{princompOne}
\alias{princompOne}
\title{First axis of a principal components analysis}
\usage{
princompOne(x, ...)
}
\arguments{
\item{x}{A matrix-like object}

\item{...}{Arguments to pass on to \code{\link{princomp}}}
}
\value{
The first axis of a PCA
}
\description{
A wrapper for \code{\link{princomp}}
}

