% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.fingerprint.regression}
\alias{print.fingerprint.regression}
\title{#' Print a fingerprint.regression (...by summarising it...)}
\usage{
\method{print}{fingerprint.regression}(x, ...)
}
\description{
#' Print a fingerprint.regression (...by summarising it...)
}

