% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.phy.structure}
\alias{print.phy.structure}
\title{Print a phylogenetic structure object}
\usage{
\method{print}{phy.structure}(x, ...)
}
\arguments{
\item{x}{\code{phy.structure} object to be printed}

\item{...}{additional arguments to be passed to
\code{summary.phy.structure} (currently there are none)}
}
\description{
Print a phylogenetic structure object
}

