% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr.R
\name{pwr}
\alias{pwr}
\alias{coef.pwr}
\alias{plot.pwr}
\alias{pwr.cv}
\title{Phylogenetically weighted regression: a method for modeling
non-stationarity on evolutionary trees}
\usage{
pwr(formula, phy4d, which.tips, bw, wfun, verbose = FALSE)

\method{coef}{pwr}(object, ...)

pwr.cv(formula, phy4d, holdout, coef = c("formula", "slope", "all"),
  model = c("pwr", "pgls"), wfun, method)

\method{plot}{pwr}(x, ...)
}
\arguments{
\item{formula}{Model formula to be fit}

\item{phy4d}{\code{phylobase} format object to be used for fitting}

\item{which.tips}{vector of integers indicating position of tips on
the tree to be used, defaults to all species on the tree}

\item{bw}{Bandwidth for model fit}

\item{wfun}{Weighting function to be used; myust be one of
\code{brownian}, \code{martins}, \code{gaussian}, or
\code{Gaus}}

\item{verbose}{Whether to be verbose during model fitting (default
\code{FALSE})}

\item{object}{\code{pwr} object to have its coefficients extracted}

\item{...}{Optional argument passed on to internal \code{phylobase}
plotting functions}

\item{holdout}{vector of integers indicating position of tips on
the tree to be held out during model fitting}

\item{coef}{coefficients to extract from the model; must be one of
\code{formula}, \code{slope}, or \code{all}}

\item{model}{Model type to be fit; must be either \code{pwr}, or
\code{pgls}}

\item{method}{for estimating weighting; should be one of subplex,
optimize, L-BFGS-B,}

\item{x}{output from}
}
\value{
Phylogenetically weighted model object; use \code{coef.pwr}
    to extract coefficients, and \code{pwr.tp} and
    \code{pwr.phylobubbles} for plotting.
}
\description{
\code{pwr.cv} performs model cross validation on the holdout set
\code{holdout}, K-fold cross validation can be specified using
\code{fold}. \code{pwr.treePlot} plots the phylogeny and allows for
plotting of PWR model coefficients at the tips of the tree. PWR
model coefficients to be pllotted must be appended to the phylo4d
object. \code{pwr.phylobubbles} plots either circles or squares
corresponding to the magnitude of each cell of a phylo4d object,
and allows the plotting of PWR moel coefficients.  Model
coefficients to be plotted must be appended to the phylo4d object
}
\note{
Some information about this
}
\examples{
# Below are some examples in order of increasing complexity: (1) a
# simple demonstration, and (2) some cross-validation examples.

##########################################
# (1) A simple example ###################
##########################################
# Make up some data
# - Note the deep bifurcation in the slope term
if(FALSE){
tree <- read.tree(text=
    "(((A:1,B:1):1,(C:1,D:1):1):1,((E:1,F:1):1,(G:1,H:1):1):1);")
dat <- data.frame(
   x = c(0, 1, 2, 3, 0, 1, 2, 3),
   y = rnorm(8, c(1, 2, 3, 4, 2, 1.5, 1, 0.5), sd=0.25)
)

# Combine in phylobase format, and then run a scatterplot
c.data <- phylo4d(tree, dat)
plot(tipData(c.data)$x, tipData(c.data)$y, pch=rep(c(1,16), each=4),
    xlab="Predictor", ylab="Response", bty="l")
text(tipData(c.data)$x, tipData(c.data)$y, labels=rownames(tipData(c.data)),
    pos=3, cex=0.8)

# Phylogenetically weighted-regression with various weighting functions
pwr.b <- coef(pwr(y ~ x, combined.data, wfun="brownian"))
pwr.m <- coef(pwr(y ~ x, combined.data, wfun="martins"))
pwr.g <- coef(pwr(y ~ x, combined.data, wfun="gaussian"))
pwr.G <- coef(pwr(y ~ x, combined.data, wfun="Gauss"))

# Run a PGLS
pgls <- gls(y ~ x, data=dat, correlation=corBrownian(phy=tree))

# Contrast the PWR and PGLS estimates
pwr.treePlot(addData(c.data, data.frame(gest=coef(pgls)[2], glb=confint(pgls)[2,1],
   gub=confint(pgls)[2,2], pwr.m))[,c("est", "lb", "ub", "gest", "glb",
   "gub")], show.tip.label=TRUE, lower=-1., upper=1.5, pex=2, aex=3.5)
}
}
\references{
Davies, T. J., Regetz, J., Wolkovich, E. M., & McGill,
    B. J. (2019). Phylogenetically weighted regression: A method
    for modelling non-stationarity on evolutionary trees. Global
    Ecology and Biogeography, 28(2), 275-285.
}
\seealso{
caper:pgls phylobase:tp phylobase:phylobubbles
}
\author{
Jim Regetz, T. Jonathan Davies, Elizabeth M. Wolkovich,
    Brian J. McGill, and a tiny bit of cosmetic tidying from Will
    Pearse. Plotting code (\code{pwr.tp} and
    \code{pwr.phylobubbles}) only slightly adapted from code in
    \code{phylobase}; thank you to the authors of \code{phylobase}!
}
