% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shape}
\alias{shape}
\title{Calculate dissimilarity metrics across communities}
\usage{
shape(data, metric = c("all", "psv", "psr", "mpd", "pd", "colless", "gamma",
  "taxon", "eigen.sum", "cadotte.pd", "mfpd"), which.eigen = 1,
  phylogeneticWeight = 0.5, removeErrors = TRUE)
}
\arguments{
\item{data}{a comparative community ecology object}

\item{metric}{specify (a) particular metric(s) to calculate
(unifrac, pcd, phylosor), or the default 'all'}

\item{which.eigen}{TODO}

\item{phylogeneticWeight}{phylogenetic weighting parameter for
computing a functional phylogenetic distance matrix}

\item{removeErrors}{If \code{FALSE}, metrics that failed to be
computed are returned as error messages}
}
\value{
cc.dissimilarity object (a named list with the output from
each metric)
}
\description{
\code{dissimilarity} calculates dissimilarity metrics in
comparative.comm communities
FIXME: properly describe shape indices, not dissimilarity (right?)
}
\details{
This calculates UniFrac, PCD (and its subcomponents), and
the phylogenetic Sorenson's index, all defined as dissimilarity
metrics in Pearse et al.
}
\note{
This function uses XXX's version of the PCD function, not
that included in picante
}
\examples{
\dontrun{
data(phylocom)
data <- comparative.comm(phylocom$phy, phylocom$sample)
shape(data)
data(laja)
data <- comparative.comm(invert.tree, river.sites, invert.traits)
(output<-shape(data))
str(output)
shape(data, "colless")
shape(data, "eigen.sum", which.eigen = 2)
}
}
\author{
Matt Helmus, Will Pearse
}
\references{
Pearse W.D., Purvis A., Cavender-Bares J. & Helmus
M.R. (2014). Metrics and Models of Community Phylogenetics. In:
Modern Phylogenetic Comparative Methods and Their Application in
Evolutionary Biology. Springer Berlin Heidelberg, pp. 451-464.

\code{PSV,PSR} Helmus M.R., Bland T.J., Williams C.K. &
Ives A.R. (2007). Phylogenetic measures of biodiversity. American
Naturalist, 169, E68-E83.

\code{PD} Faith D.P. (1992). Conservation evaluation
and phylogenetic diversity. Biological Conservation, 61, 1-10.

\code{colless} Colless D.H. (1982). Review of
phylogenetics: the theory and practice of phylogenetic
systematics. Systematic Zoology, 31, 100: 104.
}

