% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.bd.tr.tree}
\alias{sim.bd.tr.tree}
\title{\code{sim.bd.tr.tree} simulate phylogeny and trait under birth/death process linked to trait evolution}
\usage{
sim.bd.tr.tree(speciate = 0.1, extinction = 0.025, time.steps = 20,
  tr.range = c(0, 1), sp.tr = 2, ext.tr = 1, tr.walk = 0.2,
  tr.wrap = TRUE)
}
\arguments{
\item{speciate}{probability each species will speciate in each
time-step (0-1)}

\item{extinction}{probability each species will go extinct in each
time-step (0-1)}

\item{time.steps}{number of time-steps for simulation}

\item{tr.range}{vector of length two specifying boundaries for
trait values (see notes); initial two species will be at the 25th
and 75th percentiles of this space (which is also the
boundary). See also \code{tr.wrap}}

\item{sp.tr}{speciation rate's interaction with the minimum
distance between a species and the species most similar to it (see
details)}

\item{ext.tr}{extinction rate's interaction with the minimum
distance between a species and the species most similar to it (see
details)}

\item{tr.walk}{at each time-step a species not undergoing
speciation or extinction has its trait value drawn from a
distribution centered at its current value and with a standard
deviation set by this value. I.e., this is the rate of the Brownian
motion trait evolution.}

\item{tr.wrap}{whether to force species' trait values to stay
within the boundary defined by \code{tr.range}; default TRUE.}
}
\value{
ape::phylo object with random tip.labels and traits that correspond to extant species
}
\description{
\code{sim.bd.tr.tree} simulate phylogeny and trait under birth/death process linked to trait evolution
}
\details{
Simuates a phylogeny under the above parameters; an
extension of \code{sim.bd.tree} and all the caveats of that
function apply here. The extension is to include the evolution of a
trait under Brownain motion (\code{tr.walk}; but see below), and to
allow alteration of speciation and extinction based on this
trait. Species have greater or lesser rates of
speciation/extinction depending on whether they have a trait value
similar to other species (\code{sp.tr}, \code{ext.tr}). When a
speciation event happens, the two daughters split evenly about the
ancestor's trait value, taking values half-way to whatever the
nearest specie's value is.
}
\author{
Will Pearse
}

