% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.bd.tree}
\alias{sim.bd.tree}
\title{\code{sim.bd.tree} simulate phylogeny under birth/death process}
\usage{
sim.bd.tree(speciate = 0.1, extinction = 0.025, time.steps = 20)
}
\arguments{
\item{speciate}{probability each species will speciate in each
time-step (0-1)}

\item{extinction}{probability each species will go extinct in each
time-step (0-1)}

\item{time.steps}{number of time-steps for simulation}
}
\value{
ape::phylo object with random tip.labels
}
\description{
\code{sim.bd.tree} simulate phylogeny under birth/death process
}
\details{
Simuates a phylogeny under the above parameters. There are
two important things to note: (1) speciation is randomised before
extinction, and only one thing can happen to a lineage per timestep
- thus this isn't a fair randomisation, and (2) I'm fairly certain
this code works fine, but particularly weird parameter values could
well cause a crash.
}
\examples{
\dontrun{
tree <- sim.bd.tree(0.1, 0, 10)
plot(tree)
}
}
\author{
Will Pearse
}

