% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.meta.comm}
\alias{sim.meta.comm}
\title{\code{sim.meta.comm} simulate a (sort of) meta-community}
\usage{
sim.meta.comm(size = 10, n.spp = 8, timesteps = 10, p.migrate = 0.05,
  env.lam = 10, abund.lam = 5, stoch.lam = 1)
}
\arguments{
\item{size}{the length and width of the meta-community in grid cells}

\item{n.spp}{number of species}

\item{timesteps}{number of time-steps (each discrete)}

\item{p.migrate}{probability that a group of species in each grid
cell will migrate to another grid cell each timestep (i.e., 10
cells occuped by species A --> 10*p.migrate chance of migration)}

\item{env.lam}{Lambda value for Poisson distribution used to
distribute environmental quality; essentially the carrying capacity
(for each species separately) for that cell}

\item{abund.lam}{Lambda value for Poisson distribution used to
distribute initial abundances and abundance after migration}

\item{stoch.lam}{Lambda value for Poisson distribution of noise
added to the next step abundance calculation. With equal chance,
this is taken as either a positive or a negative number (see
details if you're confused as to why this is Poisson!)}
}
\value{
List with the species abundances (as a 3D array) and the
environmental quality (carrying capacities)
}
\description{
\code{sim.meta.comm} simulate a (sort of) meta-community
}
\details{
Simulates species moving through a metacommunity. At each
time-step each cell's next abundance for each species is
env.quality - current.abundance + stochastic, and a species gets as
many chances to migrate in each time-step as it has cells (the same
cell could migrate multiple times). I use a Poisson for everything
because I don't want half-species (these are individuals, because I
eventually want to make a phylogeny of these), and keeping
everything in Poisson makes it easier to compare the relative rates
of everything.
}
\author{
Will Pearse
}

