\name{[.comparative.comm}
\alias{[.comparative.comm}
\title{Slice out the species or site that you want from your comparative.comm object}
\usage{
\method{[}{comparative.comm}(x, sites, spp, warn = FALSE)
}
\description{
Slice out the species or site that you want from your
comparative.comm object
}

