\name{summary.eco.trait.regression}
\alias{summary.eco.trait.regression}
\title{\code{summary.eco.trait.regression} summarises an eco.trait.regression regression}
\usage{
\method{summary}{eco.trait.regression}(object, ...)
}
\description{
\code{summary.eco.trait.regression} summarises an
eco.trait.regression regression
}

