% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.fingerprint.regression}
\alias{summary.fingerprint.regression}
\title{Summarise a fingerprint.regression}
\usage{
\method{summary}{fingerprint.regression}(object, ...)
}
\arguments{
\item{object}{\code{fingerprint.regression} object}

\item{...}{ignored}
}
\description{
Summarise a fingerprint.regression
}

