% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.phy.structure}
\alias{summary.phy.structure}
\title{Summarise phylogenetic structure objects}
\usage{
\method{summary}{phy.structure}(object, ...)
}
\arguments{
\item{object}{\code{phy.structure} object to be summarised}

\item{...}{additional arguments (currently there are none)}
}
\description{
Summarise phylogenetic structure objects
}

