% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/traits.R
\name{traitgram.cc}
\alias{princompOne}
\alias{traitgram.cc}
\title{Traitgram for comparative community object}
\usage{
traitgram.cc(object, trait, moreArgs = NULL, ...)

princompOne(x, ...)
}
\arguments{
\item{object}{A \code{\link{comparative.comm}} object.}

\item{trait}{Which trait to plot.  If \code{\link{missing}}, use
the first trait.  If a positive \code{\link{numeric}} vector of
\code{\link{length}} one, use the \code{as.integer(trait)}th
trait.  If a \code{\link{numeric}} vector, use it instead of the
trait data in \code{object}.  If a \code{\link{character}} vector
of \code{\link{length}} one, use the trait with that name.  If a
\code{\link{function}} pass the trait data frame through that
function and use the result (\code{\link{princompOne}} is a
wrapper).  If an \code{\link{expression}}, evaluate that
expression in the environment of the trait data and use the
result.  If a \code{\link{character}} vector, then convert to an
expression and evaluate in the environment of the trait data and
use the result.}

\item{moreArgs}{List of more arguments to pass on to \code{trait}
(if its a \code{\link{function}}).}

\item{...}{Additional arguments to be passed on to
\code{\link[picante:traitgram]{traitgram}} or \code{prcomp} for
\code{traitgram.cc} and \code{princomOne} respectively.}

\item{x}{A matrix-like object}
}
\value{
\code{traitgram.cc}: see
\code{\link[picante:traitgram]{traitgram}}

\code{princompOne}: the first axis of a PCA
}
\description{
\code{traitgram.cc} A wrapper for the
\code{\link[picante:traitgram]{traitgram}} function in the
\code{picante} package.

\code{princompOne} A very soft wrapper for \code{\link{princomp}}
}

