% Generated by roxygen2 (4.0.1): do not edit by hand
\name{traits.dist}
\alias{dist.func.default}
\alias{traits.dist}
\alias{traits.dist.comparative.comm}
\alias{traits.dist.data.frame}
\alias{traits.dist.default}
\title{Make trait distance matrices}
\usage{
traits.dist(x, dist.func = dist.func.default, ...)

\method{traits.dist}{comparative.comm}(x, dist.func = dist.func.default,
  alltogether = TRUE, ...)

\method{traits.dist}{default}(x, dist.func = dist.func.default, ...)

\method{traits.dist}{data.frame}(x, dist.func = dist.func.default, ...)

dist.func.default(x)
}
\arguments{
\item{x}{an object}

\item{dist.func}{a function for computing distances.  The default,
\code{dist.func.default}, returns a Euclidean distance of the
scaled and centred data.}

\item{alltogether}{should one multivariate distance matrix be
computed for all traits at once (\code{alltogether = TRUE}) or for
each trait at a time (\code{alltogether = FALSE})?}

\item{...}{not used}
}
\description{
Make trait distance matrices
}
\seealso{
Other distances: \code{\link{comm.dist}},
  \code{\link{comm.dist.comparative.comm}},
  \code{\link{comm.dist.matrix}};
  \code{\link{funct.phylo.dist}}; \code{\link{phylo.dist}},
  \code{\link{phylo.dist.comparative.comm}},
  \code{\link{phylo.dist.phylo}};
  \code{\link{pianka.dist}},
  \code{\link{pianka.dist.comparative.comm}},
  \code{\link{pianka.dist.matrix}}
}

