% Generated by roxygen2 (4.0.1): do not edit by hand
\name{traitNames}
\alias{envNames}
\alias{traitNames}
\title{Trait and environmental variable names}
\usage{
traitNames(object)

envNames(object)
}
\arguments{
\item{object}{A \code{\link{comparative.comm}} object}
}
\value{
Names of the traits or environmental variables
}
\description{
Trait and environmental variable names
}

