% Generated by roxygen2 (4.0.1): do not edit by hand
\name{comm.dist}
\alias{comm.dist}
\alias{comm.dist.comparative.comm}
\alias{comm.dist.matrix}
\title{Make ecological co-existence matrices}
\usage{
comm.dist(x)

\method{comm.dist}{matrix}(x)

\method{comm.dist}{comparative.comm}(x)
}
\arguments{
\item{x}{an object}
}
\description{
Make ecological co-existence matrices
}
\seealso{
Other distances: \code{\link{dist.func.default}},
  \code{\link{traits.dist}},
  \code{\link{traits.dist.comparative.comm}},
  \code{\link{traits.dist.data.frame}},
  \code{\link{traits.dist.default}};
  \code{\link{funct.phylo.dist}}; \code{\link{phylo.dist}},
  \code{\link{phylo.dist.comparative.comm}},
  \code{\link{phylo.dist.phylo}}
}

