% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dispersion}
\alias{dispersion}
\title{Calculate dispersion metrics across communities}
\usage{
dispersion(data, metric = c("all", "sesmpd", "sesmntd", "sespd", "innd", "d"),
  permute = 1000)
}
\arguments{
\item{data}{a \code{comparative.comm} object}

\item{permute}{the number of null permutations to perform}

\item{metric}{specify particular metrics to calculate, default is \code{all}}
}
\value{
cc.dispersion object (a named list with the output from
each metric)
}
\description{
\code{dispersion} calculates dispersion metrics in comparative.comm communities
}
\details{
Calculates various metrics of phylogenetic biodiversity
that are categorized as \emph{dispersion} metrics by Pearse
\emph{et al.} (2014). All these are defined as dispersion metrics
in Pearse et al., Dc=0 is the Brownian expectation, Dc=1 is the
random expectation.
}
\examples{
\dontrun{
data(phylocom)
data(laja)
data <- comparative.comm(invert.tree, river.sites, invert.traits)
dispersion(data)
dispersion(data, 100, "sesmpd")
}
}
\author{
Matt Helmus, Will Pearse
}

