% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{drop_tip}
\alias{drop_tip}
\title{Trim a phylogeny}
\usage{
drop_tip(tree, spp)
}
\arguments{
\item{tree}{An \code{\link[ape:phylo]{phylo}} object}

\item{spp}{A vector of species (one, many, or none) to be removed
from \code{tree}}
}
\value{
\code{\link[ape:phylo]{phylo}} object
}
\description{
This is a weak wrapper around \code{ape}'s
\code{\link{drop.tip}}. Importantly, if asked to drop no species
from a phylogeny, it will just return the phylogeny (not an empty
phylogeny, as \code{\link{drop.tip}}) will.
}
\seealso{
\code{\link[ape:drop.tip]{drop.tip}} \code{\link[ape:extract.clade]{extract.clade}}
}

