% Generated by roxygen2 (4.0.1): do not edit by hand
\name{evenness}
\alias{evenness}
\title{Calculate evenness phylogenetic biodiversity metrics across communities}
\usage{
evenness(data, metric = c("all-quick", "all", "rao", "taxon", "entropy",
  "cadotte", "lambda", "delta", "kappa", "mpd", "mntd", "pse", "dist.fd"),
  sqrt.phy = FALSE, traitgram = NULL, traitgram.p = 2, ext.dist = NULL,
  ...)
}
\arguments{
\item{data}{a comparative community ecology object}

\item{metric}{specify particular metrics to calculate, default is
\code{all} (\code{all-quick} will calculate everything bar the
Pagel metrics and Rao's D, which can be slow) #'}

\item{sqrt.phy}{If
TRUE (default is FALSE) your phylogenetic distance matrix will be
square-rooted; specifying TRUE will force the square-root
transformation on phylogenetic distance matrices (in the spirit of
Leitten and Cornwell, 2014). See `details' for details about
different metric calculations when a distance matrix is used.}

\item{sqrt.phy}{If TRUE (default is FALSE) your phylogenetic
distance matrix will be square-rooted; specifying TRUE will force
the square-root transformation on phylogenetic distance matrices
(in the spirit of Leitten and Cornwell, 2014). See `details' for
details about different metric calculations when a distance matrix
is used.}

\item{traitgram}{If not NULL (default), a number to be passed to
\code{funct.phylo.dist} (\code{phyloWeight}; the `a' parameter),
causing analysis on a distance matrix reflecting both traits and
phylogeny (0-->only phylogeny, 1--> only traits; see
\code{funct.phylo.dist}). If a vector of numbers is given,
\code{shape} iterates across them and returns a \code{data.frame}
with coefficients from each iteration. See `details' for details
about different metric calculations when a distance matrix is used.}

\item{traitgram.p}{A value for `p' to be used in conjunction with
\code{traitgram} when calling \code{funct.phylo.dist}.}

\item{...}{Additional arguments to passed to metric functions (unlikely you will want this!)}
}
\value{
a \code{phy.structure} list object of metric values
}
\description{
\code{evenness} calculates phylogenetic biodiversity metrics
}
\details{
Calculates various metrics of phylogenetic biodiversity
that are categorized as \emph{evenness} metrics by Pearse \emph{et
al.} (2014)
}
\examples{
\dontrun{
data(laja)
data <- comparative.comm(invert.tree, river.sites, invert.traits)
evenness(data)
evenness(data, "rao")
}
}
\author{
M.R. Helmus, Will Pearse
}
\references{
Pearse W.D., Purvis A., Cavender-Bares J. & Helmus
M.R. (2014). Metrics and Models of Community Phylogenetics. In:
Modern Phylogenetic Comparative Methods and Their Application in
Evolutionary Biology. Springer Berlin Heidelberg, pp. 451-464.

\code{pse} Helmus M.R., Bland T.J., Williams C.K. &
Ives A.R. (2007). Phylogenetic measures of biodiversity. American
Naturalist, 169, E68-E83.

Pearse W.D., Purvis A., Cavender-Bares J. & Helmus M.R. (2014). Metrics and Models of Community Phylogenetics. In: Modern Phylogenetic Comparative Methods and Their Application in Evolutionary Biology. Springer Berlin Heidelberg, pp. 451-464.

\code{pse} Helmus M.R., Bland T.J., Williams C.K. & Ives A.R. (2007). Phylogenetic measures of biodiversity. American Naturalist, 169, E68-E83.

\code{rao} Webb C.O. (2000). Exploring the phylogenetic structure of ecological communities: An example for rain forest trees. American Naturalist, 156, 145-155.

\code{taxon} Clarke K.R. & Warwick R.M. (1998). A taxonomic distinctness index and its statistical properties. J. Appl. Ecol., 35, 523-531.

\code{entropy} Allen B., Kon M. & Bar-Yam Y. (2009). A New Phylogenetic Diversity Measure Generalizing the Shannon Index and Its Application to Phyllostomid Bats. The American Naturalist, 174, 236-243.

\code{cadotte} (i.e., \emph{PAE, IAC, Haed, Eaed}) Cadotte M.W., Davies T.J., Regetz J., Kembel S.W., Cleland E. & Oakley T.H. (2010). Phylogenetic diversity metrics for ecological communities: integrating species richness, abundance and evolutionary history. Ecology Letters, 13, 96-105.

\code{lambda}

\code{delta}

\code{kappa}
}

