\name{pez-internal}
\alias{.Random.seed}
\alias{.eco.evo.clade.regression}
\alias{.eco.evo.regression}
\alias{.eco.null}
\alias{.eco.phy.regression}
\alias{.plot.regression}
\alias{.prepare.regression.output}
\alias{.summary.regression}
\alias{summary.phy.structure}
\title{Internal pez Functions}
\alias{.removeErrors}
\description{
 Internal pez functions
}
\details{


  These are not to be called by the user.}
  \keyword{internal}