% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phylo.dist}
\alias{phylo.dist}
\alias{phylo.dist.comparative.comm}
\alias{phylo.dist.phylo}
\title{Make phylogenetic distance matrices}
\usage{
phylo.dist(x, ...)

\method{phylo.dist}{phylo}(x, ...)

\method{phylo.dist}{comparative.comm}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{not used}
}
\description{
Make phylogenetic distance matrices
}
\seealso{
Other distances: \code{\link{comm.dist}},
  \code{\link{comm.dist.comparative.comm}},
  \code{\link{comm.dist.matrix}};
  \code{\link{dist.func.default}},
  \code{\link{traits.dist}},
  \code{\link{traits.dist.comparative.comm}},
  \code{\link{traits.dist.data.frame}},
  \code{\link{traits.dist.default}};
  \code{\link{funct.phylo.dist}};
  \code{\link{pianka.dist}},
  \code{\link{pianka.dist.comparative.comm}},
  \code{\link{pianka.dist.matrix}}
}

