% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pianka.dist}
\alias{pianka.dist}
\alias{pianka.dist.comparative.comm}
\alias{pianka.dist.matrix}
\title{Make environmental tolerance distance matrices}
\usage{
pianka.dist(x, ...)

\method{pianka.dist}{matrix}(x, env = NULL, ...)

\method{pianka.dist}{comparative.comm}(x, alltogether = TRUE, ...)
}
\arguments{
\item{x}{an object}

\item{alltogether}{see \code{\link{traits.dist}}}

\item{...}{not used}
}
\description{
Make environmental tolerance distance matrices
}

