% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.fingerprint.regression}
\alias{print.fingerprint.regression}
\title{#' Print a fingerprint.regression (...by summarising it...)}
\usage{
\method{print}{fingerprint.regression}(x, ...)
}
\arguments{
\item{x}{\code{fingerprint.regression} object}

\item{...}{arguments passed to
\code{summary.fingerprint.regression} (currently ignored)}
}
\description{
#' Print a fingerprint.regression (...by summarising it...)
}

