% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shape}
\alias{shape}
\title{Calculate dissimilarity metrics across communities}
\usage{
shape(data, metric = c("all", "psv", "psr", "mpd", "pd", "colless", "gamma",
  "taxon", "eigen.sum", "cadotte.pd", "mfpd"), which.eigen = 1,
  phylogeneticWeight = 0.5, removeErrors = TRUE)
}
\arguments{
\item{data}{a comparative community ecology object}

\item{metric}{specify (a) particular metric(s) to calculate
(unifrac, pcd, phylosor), or the default 'all'}

\item{which.eigen}{TODO}

\item{phylogeneticWeight}{phylogenetic weighting parameter for
computing a functional phylogenetic distance matrix}

\item{removeErrors}{If \code{FALSE}, metrics that failed to be
computed are returned as error messages}
}
\value{
cc.dissimilarity object (a named list with the output from
each metric)
}
\description{
\code{dissimilarity} calculates dissimilarity metrics in
comparative.comm communities
FIXME: properly describe shape indices, not dissimilarity (right?)
}
\details{
This calculates UniFrac, PCD (and its subcomponents), and
the phylogenetic Sorenson's index, all defined as dissimilarity
metrics in Pearse et al.
}
\note{
This function uses XXX's version of the PCD function, not
that included in picante
}
\examples{
\dontrun{
data(phylocom)
data <- comparative.comm(phylocom$phy, phylocom$sample)
shape(data)
shape(data, "colless")
shape(data, "eigen.sum", which.eigen = 2)
}
}
\author{
Matt Helmus, Will Pearse
}

