% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shape}
\alias{shape}
\title{Calculate shape phylogenetic biodiversity metrics across communities}
\usage{
shape(data, metric = c("all", "psv", "psr", "mpd", "pd", "colless", "gamma",
  "taxon", "eigen.sum", "cadotte.pd", "mfpd"), which.eigen = 1,
  phylogeneticWeight = 0.5, removeErrors = TRUE)
}
\arguments{
\item{data}{a \code{comparative.comm} object}

\item{metric}{specify particular metrics to calculate, default is \code{all}}

\item{which.eigen}{The eigen vector to calculate for the PhyloEigen metric}
}
\value{
a \code{phy.structure} list object of metric values
}
\description{
\code{shape} calculates phylogenetic biodiversity metrics
}
\details{
Calculates various metrics of phylogenetic biodiversity that are categorized as \emph{shape} metrics by Pearse \emph{et al.} (2014)
}
\examples{
\dontrun{
data(laja)
data <- comparative.comm(invert.tree, river.sites, invert.traits)
(output<-shape(data))
str(output)
shape(data, "colless")
shape(data, "eigen.sum", 2)
}
}
\author{
M.R. Helmus, Will Pearse
}
\references{
Pearse W.D., Purvis A., Cavender-Bares J. & Helmus M.R. (2014). Metrics and Models of Community Phylogenetics. In: Modern Phylogenetic Comparative Methods and Their Application in Evolutionary Biology. Springer Berlin Heidelberg, pp. 451-464.

\code{PSV,PSR} Helmus M.R., Bland T.J., Williams C.K. & Ives A.R. (2007). Phylogenetic measures of biodiversity. American Naturalist, 169, E68-E83.

\code{PD} Faith D.P. (1992). Conservation evaluation and phylogenetic diversity. Biological Conservation, 61, 1-10.

\code{gamma}

\code{taxon} Clarke K.R. & Warwick R.M. (1998). A taxonomic distinctness index and its statistical properties. J. Appl. Ecol., 35, 523-531.

\code{eigen.sum} Diniz-Filho J.A.F., Cianciaruso M.V., Rangel T.F. & Bini L.M. (2011). Eigenvector estimation of phylogenetic and functional diversity. Functional Ecology, 25, 735-744.

\code{cadotte.pd} (i.e., \emph{Eed, Hed}) Cadotte M.W., Davies T.J., Regetz J., Kembel S.W., Cleland E. & Oakley T.H. (2010). Phylogenetic diversity metrics for ecological communities: integrating species richness, abundance and evolutionary history. Ecology Letters, 13, 96-105.
}

