% Generated by roxygen2 (4.0.1): do not edit by hand
\name{traits}
\alias{env}
\alias{traits}
\title{Trait and environmental variable names}
\usage{
traits(object)

env(object)
}
\arguments{
\item{object}{A \code{\link{comparative.comm}} object}
}
\value{
Names of the traits or environmental variables
}
\description{
Trait and environmental variable names
}

